commit message

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -18,7 +18,7 @@ int sum_them_all(const unsigned int n, ...)
 
 	for (i = 0; i < n; i++)
 	{
-		sum += va_arg(arglist, int);
+		sum += va_arg(arglist, long int);
 	}
 
 	va_end(arglist);
